check mail before send OTP
