+++ src/views/pages/Login/index.js
@@ -3,7 +3,6 @@ import ReCAPTCHA from 'react-google-recaptcha';
 import { useDispatch, useSelector } from 'react-redux';
 import { Link, useNavigate } from 'react-router-dom';
 import { yupResolver } from '@hookform/resolvers/yup';
-import * as yup from 'yup';
 
 import { signIn } from 'global/redux/auth/thunk';
 import Wrapper from 'components/Wrapper';
@@ -12,13 +11,9 @@ import Input from 'components/Input';
 import { useForm } from 'react-hook-form';
 import Button from 'components/Button/Default';
 import { capitalFirstLetter } from 'utils/helpers';
+import { signInValidate } from './validation';
 
 import './style.scss';
-
-const schema = yup.object().shape({
-  email: yup.string().email().required('Require field'),
-  password: yup.string().min(6).max(14).required('Require field'),
-});
 
 const Login = () => {
   const captchaRef = useRef(null);
@@ -31,7 +26,7 @@ const Login = () => {
     register,
     handleSubmit,
     formState: { errors },
-  } = useForm({ resolver: yupResolver(schema) });
+  } = useForm({ resolver: yupResolver(signInValidate) });
 
   const { currentUser, isLoading } = useSelector((state) => state.auth);
   const onSubmit = (account) => {
