+++ src/views/pages/Login/validation.js
@@ -1,0 +1,8 @@
+import * as yup from 'yup';
+
+const signInValidate = yup.object().shape({
+  email: yup.string().email().required('Require field'),
+  password: yup.string().min(6).max(14).required('Require field'),
+});
+
+export { signInValidate };
